Remove private src/xtpmmap.x
This file is thought top be included when iraf$pkg/xtools/ytpmmap.x is
not available. However, since version 2.14 (from 2008), this file is
available; so we can remove this.

This makes it possible to compile mscred without the sources of IRAF
itself (with a binary installation, f.e. from Debian or Fedora).